requirements.txt
# MCAT Content Hub - Extraction Pipeline Dependencies
google-generativeai>=0.8.0
PyMuPDF>=1.24.0
google-cloud-texttospeech>=2.16.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
firebase-admin>=6.5.0
python-slugify>=8.0.0
Pillow>=10.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
